#116 PutMarkLogicRecord now sends the correct data to batch_success
The bug was due to an entry being added to uriFlowFileMap based on a random UUID as opposed to the original FlowFile UUID. When the onBatchSuccess method in PutMarkLogic was called, a FlowFileInfo couldn't be found for the FlowFile UUID, resulting in an NPE. Added assertions to PutMarkLogicRecordTest to verify that the relationships are all getting the correct transfer counts. 

I don't know why the random "uri" value was being used as a key for uriFlowFileMap - this clearly seemed like a bug because every call to uriFlowFileMap involves a FlowFile UUID. 

Did a little cleanup in buildWriteEvent too - there were some unused arguments and lines of code that I removed.