Refactored QueryMarkLogic for readability
No change in functionality, just extracted a lot of methods and added some comments to improve the readability of the code. Also reordered the methods in QueryMarkLogic to make it read more top-to-bottom. 

Notably, QueryBatcherBuilder was extracted into a separate class to make it and QueryMarkLogic easier to understand. It is not intended yet for reuse outside of this project though, it's just an attempt to simplify the code a bit.