DEVEXP-44: DeleteML includes attributes from incoming FlowFile
Had to modify `QueryMarkLogic` in order to accomplish this and ended up fixing a good chunk of that class too, though tests will be written to ensure that in a separate ticket/PR. The main change was including `Map<String, String> incomingAttributes` in the `buildQueryBatchListener` method so that subclasses could copy the incoming FlowFile attributes into each new FlowFile that it creates. 

Also added a missing test scenario for `DeleteMarkLogic` where the user is able to read URIs but not delete them. This tests the "on batch failure" listener in `DeleteMarkLogic`.