Standardizing on use of createFlowFileWithAttributes
The intent is to make it easier in a future release to support the generation of a RECEIVE provenance event. 

This also ensures that FORK or JOIN provenance events are not created, as per the NiFi javadocs, those are not appropriate when creating new FlowFiles based on data received from an external system (MarkLogic). 

Also did some renaming to standardize on `incomingFlowFile` as a variable name.