DEVEXP-410 Now supporting cloud authentication
Improved the descriptions of connection properties too, as I realized that with "expression language" support, the allowable values are not shown by NiFi (which is odd - I would have expected an editable dropdown in the NiFi GUI). 

Tested this manually as we don't yet have the ability to run automated tests against cloud.