Bumped to 1.24.0
Updated dependencies as well.

Note that the connector will still work on older versions of NiFi, as we are not yet depending on any new features in the connector API. 

Had to rework some tests because in between NiFi 1.16 and 1.24, NiFi changed from JUnit 4 to 5 (which is good, just required some test changes).

Added "flows-for-manual-testing", which had been on our internal Wiki before.